pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weatherwithyou"
version = "0.1.0"
description = "Backend service for Weather With You."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "pydantic>=2.8,<3.0",
    "httpx>=0.27,<1.0",
    "psycopg[binary]>=3.2,<4.0",
    "python-dotenv>=1.0,<2.0",
    "google-genai>=1.30,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.5,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["weatherwithyou*"]

[tool.ruff]
target-version = "py312"
line-length = 100
